Fixed Pause Button
The problem was that as soon as the pause button detected the on up
event, It ran PauseToggle().
After that, Update() ran. It saw that the pause menu was active and that
there had been a mouseup event (from the initial pause button press)
this then unpaused the game. so we were pausing and unpausing the game
really quick.