Rollover for cutscenes
If the player waits for the line being read to end without clicking, it
will roll over to the next line after 3 seconds.

If the player clicks, we need to give them more than 3 seconds to read
the line incase it is a long one. So it will roll over after 10 seconds.